% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcBoundaries.R
\name{calc_boundaries}
\alias{calc_boundaries}
\title{calc_boundaries}
\usage{
calc_boundaries(x, cumulative_proportions)
}
\arguments{
\item{x}{[\code{matrix(x,y)}]\cr 2D matrix of data values.}

\item{cumulative_proportions}{[\code{numerical}]\cr Vector of class cumulative proportions, as generated by \code{w2cp}.}
}
\value{
Numerical vector with boundaries for matrix classifcation
}
\description{
Determine upper class boundaries for classification of a matrix with values ranging 0-1 based upon an
vector of cumulative proportions.
}
\examples{
x <- matrix(runif(100,0,1),10,10)
y <- NLMR::w2cp(c(0.5, 0.25, 0.25)) #cumulative proportion
calc_boundaries(x,y)

}
