% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcBoundaries.R
\docType{methods}
\name{calcBoundaries-method}
\alias{calcBoundaries-method}
\alias{calcBoundaries,matrix-method}
\alias{calcBoundaries}
\title{Method calcBoundaries}
\usage{
calcBoundaries(array, cumulativeProportions)

\S4method{calcBoundaries}{matrix}(array, cumulativeProportions)
}
\arguments{
\item{array}{2D array of data values}

\item{cumulativeProportions}{1D array of class cumulative proportions}
}
\value{
Rectangular matrix with values ranging from 0-1
}
\description{
Method calcBoundaries

Determine upper class boundaries for classification of an array with values ranging 0-1 based upon an
array of cumulative proportions.
}
\examples{
\dontrun{
calcBoundaries(rasterNLM)
}

}
