% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcBoundaries.R
\name{calcBoundaries}
\alias{calcBoundaries}
\title{calcBoundaries}
\usage{
calcBoundaries(x, cumulativeProportions)
}
\arguments{
\item{x}{[\code{matrix(x,y)}]\cr 2D matrix of data values.}

\item{cumulativeProportions}{[\code{numerical}]\cr Vector of class cumulative proportions.}
}
\value{
Numerical vector with boundaries for matrix classifcation
}
\description{
Determine upper class boundaries for classification of a matrix with values ranging 0-1 based upon an
vector of cumulative proportions.
}
\examples{
x <- matrix(runif(100,0,1),10,10)
y <- c(0.5, 0.25, 0.25)
calcBoundaries(x,y)

}
