% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifyArray.R
\docType{methods}
\name{classifyArray-method}
\alias{classifyArray-method}
\alias{classifyArray,matrix-method}
\alias{classifyArray}
\title{Method classifyArray}
\usage{
classifyArray(array, weighting)

\S4method{classifyArray}{matrix}(array, weighting)
}
\arguments{
\item{array}{2D array of data values}

\item{weighting}{A list of numeric values}
}
\value{
Rectangular matrix with values ranging from 0-1
}
\description{
Method classifyArray

Classify an array with values ranging 0-1 into proportions based upon a list of class  weighting .
}
\examples{
\dontrun{
classifyArray(rasterNLM)
}

}
