% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifyMatrix.R
\name{classifyMatrix}
\alias{classifyMatrix}
\title{classifyMatrix}
\usage{
classifyMatrix(x, weighting)
}
\arguments{
\item{x}{[\code{matrix(x,y)}]\cr 2D matrix of data values.}

\item{weighting}{[\code{numerical}]\cr  Vector of numeric values.}
}
\value{
Rectangular matrix reclassified values.
}
\description{
Classify a matrix with values ranging 0-1 into proportions based upon a vector of class weightings.
}
\details{
The length of the weighting vector determines the number of classes in the resulting matrix.
}
\examples{
x <- matrix(runif(100, 0, 1), 10, 10)
y <- c(0.5, 0.25, 0.25)
classifyMatrix(x, y)


}
