% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distancegradient.R
\docType{methods}
\name{distancegradientNLM-method}
\alias{distancegradientNLM-method}
\alias{distancegradientNLM,ANY-method}
\alias{distancegradientNLM}
\title{Method distancegradientNLM}
\usage{
distancegradientNLM(nCol, nRow, origin, rescale = TRUE)

\S4method{distancegradientNLM}{ANY}(nCol, nRow, origin, rescale = TRUE)
}
\arguments{
\item{nCol}{Number of columns (numerical)}

\item{nRow}{Number of rows (numerical)}

\item{origin}{Edge coordinates of the origin of the distance measurement (Numeric vector of xmin, xmax, ymin, ymax)}

\item{rescale}{If \code{TRUE} (Standard), the values are rescaled between 0-1. Otherwise, the distance in raster units is calculated (logical)}
}
\value{
Raster with distance measurements starting in a predefined origin.
}
\description{
Method distancegradientNLM

Create a distance gradient neutral landscape model with values ranging 0-1.
}
\examples{
\dontrun{
distancegradientNLM(x = 100, y = 100, origin = (20, 30, 10, 15))
}

}
