% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distancegradient.R
\name{distancegradientNLM}
\alias{distancegradientNLM}
\title{distancegradientNLM}
\usage{
distancegradientNLM(nCol, nRow, origin, rescale = TRUE)
}
\arguments{
\item{nCol}{[\code{numerical(1)}]\cr Number of columns for the raster.}

\item{nRow}{[\code{numerical(1)}]\cr Number of rows for the raster.}

\item{origin}{[\code{numerical(4)}]\cr Edge coordinates of the origin of the distance measurement.}

\item{rescale}{[\code{logical(1)}]\cr If \code{TRUE} (default), the values are rescaled between 0-1. Otherwise, the distance in raster units is calculated.}
}
\value{
RasterLayer with distance measurements starting in a predefined origin.
}
\description{
Create a distance gradient neutral landscape model with values ranging 0-1.
}
\details{
origin is a umeric vector of xmin, xmax, ymin, ymax for a square inside the raster from which the distance is measured.
xmax and ymax must be smaller than nCol and nRow.
}
\examples{
distancegradientNLM(nCol = 100, nRow = 100, origin = c(20, 30, 10, 15))

}
