% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgegradient.R
\docType{methods}
\name{edgegradientNLM-method}
\alias{edgegradientNLM-method}
\alias{edgegradientNLM,ANY-method}
\alias{edgegradientNLM}
\title{Method edgegradientNLM}
\usage{
edgegradientNLM(nCol, nRow, direction = NA, rescale = TRUE)

\S4method{edgegradientNLM}{ANY}(nCol, nRow, direction = NA, rescale = TRUE)
}
\arguments{
\item{nCol}{Number of columns for the raster (numerical).}

\item{nRow}{Number of rows for the raster (numerical).}

\item{direction}{Direction of the gradient, if unspecified the direction is randomly determined (numeric (0-360), optional).}

\item{rescale}{If \code{TRUE} (Standard), the values are rescaled between 0-1. Otherwise, the distance in raster units is calculated (logical)}
}
\value{
Raster with random values ranging from 0-1.
}
\description{
Method edgegradientNLM

Create an edge gradient neutral landscape model with values ranging 0-1.
}
\examples{
\dontrun{
edgegradientNLM(x = 100, y = 100)
}

}
