% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgegradient.R
\name{edgegradientNLM}
\alias{edgegradientNLM}
\title{edgegradientNLM}
\usage{
edgegradientNLM(nCol, nRow, direction = NA, rescale = TRUE)
}
\arguments{
\item{nCol}{[\code{numerical(1)}]\cr Number of columns for the raster.}

\item{nRow}{[\code{numerical(1)}]\cr Number of rows for the raster.}

\item{direction}{[\code{numerical(1)}]\cr Direction of the gradient (between 0 and 360 dec), if unspecified the direction is randomly determined.}

\item{rescale}{[\code{logical(1)}]\cr If \code{TRUE} (default), the values are rescaled between 0-1.}
}
\value{
RasterLayer with xxxxxxx.
}
\description{
Create an edge gradient neutral landscape model with values ranging 0-1.
}
\examples{
edgegradientNLM(nCol = 100, nRow = 100, direction = 80)

}
