% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fracpoly.R
\name{fracpoly_best}
\alias{fracpoly_best}
\title{Best-fitting fractional polynomial}
\usage{
fracpoly_best(coef, coef_se, xmean, d = 1, pd = 0.05, method = "FE")
}
\arguments{
\item{coef}{coefficients of the localised average causal effects.}

\item{coef_se}{standard errors of the localised average causal effects.}

\item{xmean}{the mean of the exposure in each quantile.}

\item{d}{fractional polynomial degree. The default is degree 1. The other options are: 1, 2, or "both".}

\item{pd}{p-value cut-off for choosing the best-fitting fractional polynomial of degree 2 over the best-fitting fractional polynomial degree 1. This option is only used if d="both". The default is 0.05.}

\item{method}{the meta-analysis method.}
}
\value{
List of non-linear MR results from the fractional polynomial MR approach.

\item{model}{the best-fitting fractional polynomial of either 1 or 2.}

\item{p_ML}{the power of the best-fitting fractional polynomial of degree 1.}

\item{p1_ML}{the first power of the best-fitting fractional polynomial of degree 2.}

\item{p2_ML}{the second power of the best-fitting fractional polynomial of degree 2.}

\item{fp_p}{the p-value testing the best-fitting fractional polynomial of degree 1 against the linear model.}

\item{fp_d12_p}{the p-value testing the best-fitting fractional polynomial of degree 2 against the best-fitting fractional polnomial of degree 1.}

\item{d}{the degree of the best fitting fractional polynomial.}
}
\description{
fracpoly_best computes the best-fitting fractional polynomial of degrees 1 and 2.
}
\author{
James R Staley <jrstaley95@gmail.com>
}
