% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fracpoly.R
\name{fracpoly_boot}
\alias{fracpoly_boot}
\title{Fractional polynomial bootstrap}
\usage{
fracpoly_boot(y, x, g, covar, q, x0q, xcoef, family = "gaussian", xpos,
  method = "FE", nboot, d, p_ML = NULL, p1_ML = NULL, p2_ML = NULL)
}
\arguments{
\item{y}{vector of outcome values.}

\item{x}{vector of exposure values.}

\item{g}{the instrumental variable.}

\item{covar}{data.frame of covariates.}

\item{q}{the number of quantiles the exposure distribution is to be split into. Within each quantile a causal effect will be fitted, known as a localised average causal effect (LACE). The default is deciles (i.e. 10 quantiles).}

\item{x0q}{quantiles of x0 (the IV-free exposure).}

\item{xcoef}{the association between the exposure and the instrument.}

\item{family}{a description of the error distribution and link function to be used in the model. For fracpoly_mr this can be a character string naming either the gaussian (i.e. for continuous data) or binomial (i.e. for binary data) family function.}

\item{xpos}{the position used to relate x to the localised average causal effect. The default is the mean of the x-values within each quantile, otherwise specify a percentile (e.g. 0.5 corresponds to the median value).}

\item{method}{meta-regression method parsed to the rma package. The default is fixed-effects ("FE").}

\item{nboot}{the number of bootstrap replications (if required). The default is 100 replications.}

\item{d}{fractional polynomial degree. The default is degree 1. The other options are: 1, 2, or "both".}

\item{p_ML}{the power of the best-fitting fractional polynomial of degree 1.}

\item{p1_ML}{the first power of the best-fitting fractional polynomial of degree 2.}

\item{p2_ML}{the second power of the best-fitting fractional polynomial of degree 2.}
}
\value{
\item{frac_coef_boot}{a matrix of fractional polynomial results for each bootstrap replicates.}
}
\description{
fracpoly_boot computes the best-fitting fractional polynomial of degrees 1 and 2.
}
\author{
James R Staley <js16174@bristol.ac.uk>
}
