% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fracpoly.R
\name{fracpoly_figure}
\alias{fracpoly_figure}
\title{Fractional polynomial figure}
\usage{
fracpoly_figure(beta, cov, x.min, x.max, family = "gaussian", d = 1,
  p_ML = NULL, p1_ML = NULL, p2_ML = NULL, ci = "model_se",
  frac_coef_boot = NULL, ref = mean(x), pref_x = "x", pref_x_ref = "x",
  pref_y = "y", ci_type = "overall", ci_quantiles = 10, breaks = NULL)
}
\arguments{
\item{beta}{the coefficients of the best-fitting fractional polynomial.}

\item{cov}{the covariance matrix of the best-fitting fractional polynomial.}

\item{x.min}{the minimum value of the exposure.}

\item{x.max}{the maximum value of the exposure.}

\item{family}{a description of the error distribution and link function to be used in the model. For fracpoly_mr this can be a character string naming either the gaussian (i.e. "gaussian" for continuous data) or binomial (i.e. "binomial" for binary data) family function.}

\item{d}{fractional polynomial degree. The default is degree 1. The other options are: 1, 2, or "both".}

\item{p_ML}{the power of the best-fitting fractional polynomial of degree 1.}

\item{p1_ML}{the first power of the best-fitting fractional polynomial of degree 2.}

\item{p2_ML}{the second power of the best-fitting fractional polynomial of degree 2.}

\item{ci}{the type of 95\% confidence interval. There are three options: (i) using the model standard errors ("model_se"), (ii) using bootstrap standard errors ("bootstrap_se"), (iii) using bootstrap percentile confidence intervals ("bootstrap_per"). The default is the model standard errors.}

\item{frac_coef_boot}{a matrix of fractional polynomial results for each bootstrap replicates.}

\item{ref}{the reference point for the figure. This is the value of the function that represents the expected difference in the outcome compared with this reference value when the exposure is set to different values. The default is the mean of x.}

\item{pref_x}{the prefix/label for the x-axis. The default is "x".}

\item{pref_x_ref}{the prefix for the reference value displayed on the y-axis. The default is "x".}

\item{pref_y}{the prefix/label for the y-axis. The default is "y".}

\item{ci_type}{the type of confidence interval to be displayed on the graph. The default is "overall" where confidence intervals are presented as bands across the range of x. The alternative option is "quantile" where the confidence intervals are presented as error bars at the mean in each quantile of x.}

\item{ci_quantiles}{the number of quantiles at which confidence intervals are to be displayed. The default is deciles (i.e. 10).}

\item{breaks}{breaks on the y-axis of the figure.}
}
\value{
the plot of the best-fitting fractional polynomial.
}
\description{
fracpoly_figure plots the best-fitting fractional polynomial.
}
\author{
James R Staley <jrstaley95@gmail.com>
}
