% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lace.R
\name{hamardman.prod}
\alias{hamardman.prod}
\title{Hamardman product}
\usage{
hamardman.prod(coef, covar)
}
\arguments{
\item{coef}{vector of regression coefficients.}

\item{covar}{a matrix of covariates}
}
\description{
hamardman.prod computes the Hamardman product of a vector of regression coefficients and a matrix of covariates.
}
\author{
James R Staley <jrstaley95@gmail.com>
}
