% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lace.R
\name{iv_free}
\alias{iv_free}
\title{IV-free exposure}
\usage{
iv_free(y, x, g, covar, q, family = "gaussian")
}
\arguments{
\item{y}{vector of outcome values.}

\item{x}{vector of exposure values.}

\item{g}{the instrumental variable.}

\item{covar}{a matrix of covariates.}

\item{q}{the number of quantiles the exposure distribution is to be split into. Within each quantile a causal effect will be fitted, known as a localised average causal effect (LACE). The default is deciles (i.e. 10 quantiles).}

\item{family}{a description of the error distribution and link function to be used in the model (either "gaussian" or "binomial" can be specified).}
}
\value{
\item{xcoef}{the association between the exposure and the instrument.}

\item{x0}{the IV-free exposure.}

\item{x0q}{the quantiles of x0.}
}
\description{
iv_free computes the IV-free exposure.
}
\author{
James R Staley <jrstaley95@gmail.com>
}
