% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lace.R
\name{lace}
\alias{lace}
\title{Localised avergae causal effects}
\usage{
lace(y, x, g, covar = NULL, q, x0q, xc_sub = TRUE, family = "gaussian",
  xpos = "mean")
}
\arguments{
\item{y}{vector of outcome values.}

\item{x}{vector of exposure values.}

\item{g}{the instrumental variable.}

\item{covar}{data.frame of covariates.}

\item{q}{the number of quantiles the exposure distribution is to be split into. Within each quantile a causal effect will be fitted, known as a localised average causal effect (LACE). The default is deciles (i.e. 10 quantiles).}

\item{x0q}{quantiles of x0 (the IV-free exposure).}

\item{xc_sub}{compute the association between the exposure and the insturment in each quantile of x0.}

\item{family}{a description of the error distribution and link function to be used in the model (either gaussian or binomial can be specified).}

\item{xpos}{the position used to relate x to the localised average causal effect. The default is the mean of the x-values within each quantile, otherwise specify a percentile (e.g. 0.5 corresponds to the median value).}
}
\value{
List of results for the localised average causal effects for quantile.

\item{coef}{the localised average causal effect in each quantile.}

\item{coef_se}{the IV-free exposure.}

\item{xmean}{the mean of the exposure in each quantile.}

\item{xcoef_sub}{the association between the exposure and the instrument in each quantile.}

\item{xcoef_sub_se}{the standard error of the association between the exposure and the instrument in each quantile}
}
\description{
lace computes the localised average causal effect for quantile.
}
\author{
James R Staley <js16174@bristol.ac.uk>
}
