% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsq_marco.R
\name{marco_mpd}
\alias{marco_mpd}
\title{Mid-point-displacement algorithm}
\usage{
marco_mpd(n, roughness = 0.5)
}
\arguments{
\item{n}{Size of the landscape (2^n + 1)}

\item{roughness}{Degree of spatial autocorrelation}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
\itemize{
  \item The diamond step: Taking a square of four points, generate a random value at the square midpoint, where the two diagonals meet. The midpoint value is calculated by averaging the four corner values, plus a random amount. This gives you diamonds when you have multiple squares arranged in a grid.
  \item The square step: Taking each diamond of four points, generate a random value at the center of the diamond. Calculate the midpoint value by averaging the corner values, plus a random amount generated in the same range as used for the diamond step. This gives you squares again.
}
}
\examples{
\dontrun{
marco_mpd(50, 50, 0.4)
}
}
