% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric_area.R
\name{metric_area}
\alias{metric_area}
\title{metric_area}
\usage{
metric_area(nlm, poi = NULL)
}
\arguments{
\item{nlm}{[\code{matrix(x,y)}]\cr 2D matrix of data values.}

\item{poi}{[\code{numerical}]\cr  Vector of numeric values.}
}
\value{
Rectangular matrix reclassified values.
}
\description{
Classify a matrix with values ranging 0-1 into proportions based upon a vector of class weightings.
}
\details{
The length of the weighting vector determines the number of classes in the resulting matrix.
}
