% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlm_polylands.R
\name{nlm_polylands}
\alias{nlm_polylands}
\title{nlm_polylands}
\usage{
nlm_polylands(nCol, nRow, resolution = 1, option = 1, germs, g, R,
  patch_classes)
}
\arguments{
\item{nCol}{[\code{numerical(1)}]\cr
Number of columns for the raster.}

\item{nRow}{[\code{numerical(1)}]\cr
Number of rows for the raster.}

\item{resolution}{[\code{numerical(1)}]\cr
Resolution of the raster.}

\item{option}{[\code{numerical(1)}]\cr
If \code{1} (default), the Tessellation method is used to simulate the NLM.
If \code{2}, the Gibbs algorithm method is used to simulate the NLM.}

\item{germs}{[\code{numerical(1)}]\cr
Intensity parameter (a positive number)}

\item{g}{[\code{numerical(1)}]\cr
Interaction parameter (a number between 0 and 1, inclusive - only used when \code{option = 2}).}

\item{R}{[\code{numerical(1)}]\cr
Interaction radius (a non-negative number).}

\item{patch_classes}{[\code{numerical(1)}]\cr
Number of classes for germs}
}
\value{
RasterLayer
}
\description{
Simulate a spatially random neutral landscape model with values
drawn a uniform distribution.
}
\details{
The function takes the number of columns and rows as input and creates a
RasterLayer with the same extent. Each raster cell is randomly assigned a
value between 0 and 1 drawn from an uniform distribution (\code{runif(1,0,1)}).
}
\examples{
nlm_polylands(nCol = 50, nRow = 50, germs = 20)

}
\references{
Gaucherel, C. (2008) Neutral models for polygonal landscapes with linear
networks. \emph{Ecological Modelling}, 219, 39 - 48.
}
