% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{NLMR}
\alias{NLMR}
\alias{NLMR-package}
\title{NLMR}
\description{
\emph{NLMR} is an R package for simulating neutral landscape models (NLM).
By using
}
\details{
This package contains vignettes that introduce NLM and basic usage
of the \emph{NLMR} package. The supplemental
\href{https://github.com/prioritizr/nlmrdata}{\emph{nlmrdata}}
package contains a comprehensive database of simulated NLM. The vignettes
in this package are listed below.

\describe{
\item{\href{https://marcosci.github.io/NLMR/articles/background.html}{NLM Concepts + Terminology}}{Background information
 on the concepts and terminology that underpin the simulation of neutral landscape models.}
\item{\href{https://marcosci.github.io/NLMR/articles/bestiary.html}{NLMR Bestiary}}{Bestiary of the available neutral landscape models.}
\item{\href{https://marcosci.github.io/NLMR/articles/BasicUsageVignette.html}{Quickstart Guide}}{Short walk-through of the
  \emph{NLMR} package and how to handle the simulations.}
}
}
\references{
Gardner RH, Milne BT, Turnei MG, O’Neill R V. 1987. Neutral models for the analysis of broad-scale landscape pattern. \emph{Landscape Ecology} 1:19–28.
}
