% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlmr_plot.R
\docType{methods}
\name{nlmr_plot-method}
\alias{nlmr_plot-method}
\alias{nlmr_plot,RasterLayer-method}
\alias{nlmr_plot}
\title{Method nlmr_plot}
\usage{
nlmr_plot(nlm_obj)

\S4method{nlmr_plot}{RasterLayer}(nlm_obj)
}
\arguments{
\item{nlm_obj}{Raster* object}
}
\value{
rasterVis visualization
}
\description{
Method nlmr_plot

Create a random rectangular cluster neutral landscape model with values ranging 0-1.
}
\examples{
\dontrun{
nlmr_plot(nlm_obj)
}

}
