% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piecewise.R
\name{piecewise_figure}
\alias{piecewise_figure}
\title{Piecewise linear figure}
\usage{
piecewise_figure(y, x, g, covar = NULL, q = 10, xcoef, coef, x0q,
  family = "gaussian", nboot = 100, ref = mean(x), pref_x = "x",
  pref_x_ref = "x", pref_y = "y", ci_quantiles = 10, breaks = NULL)
}
\arguments{
\item{y}{vector of outcome values.}

\item{x}{vector of exposure values.}

\item{g}{the instrumental variable.}

\item{covar}{data.frame of covariates.}

\item{q}{the number of quantiles the exposure distribution is to be split into. Within each quantile a causal effect will be fitted, known as a localised average causal effect (LACE). The default is deciles (i.e. 10 quantiles).}

\item{xcoef}{the association between the exposure and the instrument.}

\item{coef}{the coefficients of the localised causal effects.}

\item{x0q}{quantiles of x0 (the IV-free exposure).}

\item{family}{a description of the error distribution and link function to be used in the model. For frac_poly_mr this can be a character string naming either the gaussian (i.e. "gaussian" for continuous data) or binomial (i.e. "binomial" for binary data) family function.}

\item{nboot}{the number of bootstrap replications (if required). The default is 100 replications.}

\item{ref}{the reference point for the figure. This is the value of the function that represents the expected difference in the outcome compared with this reference value when the exposure is set to different values. The default is the mean of x.}

\item{pref_x}{the prefix/label for the x-axis. The default is "x".}

\item{pref_x_ref}{the prefix for the reference value displayed on the y-axis. The default is "x".}

\item{pref_y}{the prefix/label for the y-axis. The default is "y".}

\item{ci_quantiles}{the number of quantiles at which confidence intervals are to be displayed. The default is deciles (i.e. 10).}

\item{breaks}{breaks on the y-axis of the figure.}
}
\value{
the plot of the piecewise linear function.
}
\description{
piecewise_figure plots the piecewise linear function.
}
\author{
James R Staley <jrstaley95@gmail.com>
}
