% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piecewise.R
\name{piecewise_mr}
\alias{piecewise_mr}
\title{Piecewise linear Mendelian randomization}
\usage{
piecewise_mr(
  y,
  x,
  g,
  covar = NULL,
  family = "gaussian",
  q = 10,
  xpos = "mean",
  nboot = 100,
  fig = TRUE,
  ref = mean(x),
  pref_x = "x",
  pref_x_ref = "x",
  pref_y = "y",
  ci_quantiles = 10,
  breaks = NULL
)
}
\arguments{
\item{y}{vector of outcome values}

\item{x}{vector of exposure values}

\item{g}{the instrumental variable}

\item{covar}{\code{data.frame} of covariates}

\item{family}{a description of the error distribution and link function
to be used in the model and is a \code{character} string naming either the
gaussian (i.e. \code{"gaussian"} for continuous data) or binomial
(i.e. \code{"binomial"} for binary data) family function (default: \code{"gaussian"})}

\item{q}{the number of quantiles the exposure distribution is to be split
into within which a causal effect will be fitted, known as localised
average causal effects (LACE) (default: \code{10})}

\item{nboot}{the number of bootstrap replications (default: \code{100})}

\item{fig}{a \code{logical} statement as to whether the user wants the results
displayed in a figure (default: \code{TRUE})}

\item{ref}{the reference point for the figure, this is the value of the
function that represents the expected difference in the outcome compared
with this reference value when the exposure is set to different values
(default: \code{mean(x)})}

\item{pref_x}{the prefix/label for the x-axis (default: \code{"x"})}

\item{pref_x_ref}{the prefix for the reference value displayed on the y-axis
(default: \code{"x"})}

\item{pref_y}{the prefix/label for the y-axis (default: \code{"y"})}

\item{ci_quantiles}{the number of quantiles at which confidence intervals
are to be displayed (default: \code{10})}

\item{breaks}{breaks on the y-axis of the figure}
}
\value{
\code{piecewise_mr} returns a \code{list} of non-linear MR results from the
piecewise linear function MR approach:

\item{n}{number of individuals}

\item{model}{
the model specifications:
number of quantiles (\code{q}),
number of bootstrap replications performed (\code{nboot})
}

\item{coefficients}{
the regression estimates:
regression coefficients (\code{beta}),
standard errors of regression coefficients (\code{se}),
lower 95\% confidence interval (\code{lci}),
upper 95\% confidence interval (\code{uci}),
p-value (\code{pval})
}

\item{lace}{
the localised average causal effect estimate in each quantile:
regression coefficients (\code{beta}),
standard errors of regression coefficients (\code{se}),
lower 95\% confidence interval (\code{lci}),
upper 95\% confidence interval (\code{uci}),
p-value (\code{pval})
}

\item{xcoef}{
the association between the instrument and the exposure in each quantile:
regression coefficients (\code{beta}),
standard errors of regression coefficients (\code{se})
}

\item{p_tests}{
the p-value of the non-linearity tests:
p-value from the quadratic test (\code{quad}),
p-value from the Cochran Q test (\code{Q})
}

\item{p_heterogeneity}{
the p-value of heterogeneity:
p-value of the Cochran Q heterogeneity test (\code{Q}),
p-value from the trend test (\code{trend}).
}
}
\description{
\code{piecewise_mr} performs a Mendelian randomization (MR) analysis
by fitting a piecewise linear function to localised average causal effects.
}
\examples{
# IV (g), exposure (x) & outcome (y)
epsx <- rexp(10000)
u <- runif(10000, 0, 1)
g <- rbinom(10000, 2, 0.3)
epsy <- rnorm(10000)
ag <- 0.25
x <- 1 + ag * g + u + epsx
y <- 0.15 * x^2 + 0.8 * u + epsy

# Covariates (covar)
c1 <- rnorm(10000)
c2 <- rnorm(10000)
c3 <- rbinom(10000, 2, 0.33)
covar <- data.frame(c1 = c1, c2 = c2, c3 = as.factor(c3))

# Analyses
fp <- fracpoly_mr(
  y = y , x = x, g = g, covar = covar,
  family = "gaussian", q = 10, d = 1, ci = "model_se",
  fig = TRUE
)
summary(fp)
plm <- piecewise_mr(
  y = y , x = x, g = g, covar = covar,
  family = "gaussian", q = 10, nboot = 100,
  fig = TRUE
)
summary(plm)

}
\author{
James Staley \href{mailto:jrstaley95@gmail.com}{jrstaley95@gmail.com}
}
