% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/planargradient.R
\docType{methods}
\name{planargradientNLM-method}
\alias{planargradientNLM-method}
\alias{planargradientNLM,ANY-method}
\alias{planargradientNLM}
\title{Method planargradientNLM}
\usage{
planargradientNLM(nCol, nRow, direction = NA, rescale = FALSE)

\S4method{planargradientNLM}{ANY}(nCol, nRow, direction = NA,
  rescale = TRUE)
}
\arguments{
\item{nCol}{Number of columns for the raster (numerical).}

\item{nRow}{Number of rows for the raster (numerical).}

\item{direction}{Direction of the gradient, if unspecified the direction is randomly determined (numeric (0-360), optional).}

\item{rescale}{If \code{TRUE} (Standard), the values are rescaled between 0-1. Otherwise, the distance in raster units is calculated (logical)}
}
\value{
Raster with random values ranging from 0-1.
}
\description{
Method planargradientNLM

Create a planar gradient neutral landscape model with values ranging 0-1.
}
\examples{
\dontrun{
planargradientNLM(nCol = 100, nRow = 100)
}

}
