% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/planargradient.R
\name{planargradientNLM}
\alias{planargradientNLM}
\title{planargradientNLM}
\usage{
planargradientNLM(nCol, nRow, direction = NA, rescale = TRUE)
}
\arguments{
\item{nCol}{[\code{numerical(1)}]\cr Number of columns for the raster.}

\item{nRow}{[\code{numerical(1)}]\cr Number of rows for the raster.}

\item{direction}{[\code{numerical(1)}]\cr Direction of the gradient, if unspecified the direction is randomly determined.}

\item{rescale}{[\code{logical(1)}]\cr If \code{TRUE} (default), the values are rescaled between 0-1.}
}
\value{
RasterLayer with random values ranging from 0-1.
}
\description{
Create a planar gradient neutral landscape model with values ranging 0-1.
}
\examples{
planargradientNLM(nCol = 100, nRow = 100)


}
