% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fracpoly.R
\name{print.fracpoly_mr}
\alias{print.fracpoly_mr}
\title{Print fractional polynomial fits}
\usage{
\method{print}{fracpoly_mr}(x, ...)
}
\arguments{
\item{x}{an object of class "fracpoly_mr".}
}
\description{
print method for class "fracpoly_mr".
}
\author{
James R Staley <jrstaley95@gmail.com>
}
