% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piecewise.R
\name{print.piecewise_mr}
\alias{print.piecewise_mr}
\title{Print piecewise_mr}
\usage{
\method{print}{piecewise_mr}(x, ...)
}
\arguments{
\item{x}{an object of class \code{"piecewise_mr"}}
}
\description{
print method for class \code{"piecewise_mr"}.
}
\author{
James Staley \href{mailto:jrstaley95@gmail.com}{jrstaley95@gmail.com}
}
