% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piecewise.R
\name{print.summary.piecewise_mr}
\alias{print.summary.piecewise_mr}
\title{Print summary of piecewise linear fits}
\usage{
\method{print}{summary.piecewise_mr}(x, ...)
}
\arguments{
\item{x}{an object of class "piecewise_mr".}
}
\description{
print summary method for class "piecewise_mr".
}
\author{
James R Staley <jrstaley95@gmail.com>
}
