% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomCluster.R
\docType{methods}
\name{randomClusterNLM-method}
\alias{randomClusterNLM-method}
\alias{randomClusterNLM,ANY-method}
\alias{randomClusterNLM}
\title{Method randomClusterNLM}
\usage{
randomClusterNLM(nCol, nRow, n, p, rescale = TRUE)

\S4method{randomClusterNLM}{ANY}(nCol, nRow, n, p, rescale = TRUE)
}
\arguments{
\item{nCol}{Number of columns for the raster (numerical)}

\item{nRow}{Number of rows for the raster (numerical)}

\item{n}{Clusters are defined using a set of neighbourhood structures, 4 = '4-neighbourhood', 8 = '8-neighbourhood'}

\item{p}{The p value controls the proportion of elements randomly selected to form clusters.}

\item{rescale}{If \code{TRUE} (Standard), the values are rescaled between 0-1. Otherwise, the distance in raster units is calculated (logical)}
}
\value{
Raster with random values ranging from 0-1.
}
\description{
Method randomClusterNLM

Create a random rectangular cluster neutral landscape model with values ranging 0-1.
}
\examples{
\dontrun{
randomClusterNLM(nCol = 100, nRow = 100)
}

}
