% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomCluster.R
\name{randomClusterNLM}
\alias{randomClusterNLM}
\title{randomClusterNLM}
\usage{
randomClusterNLM(nCol, nRow, neighbourhood, p, rescale = TRUE)
}
\arguments{
\item{nCol}{[\code{numerical(1)}]\cr Number of columns for the raster.}

\item{nRow}{[\code{numerical(1)}]\cr Number of rows for the raster.}

\item{neighbourhood}{[\code{numerical(1)}]\cr Clusters are defined using a set of neighbourhood structures, 4 = '4-neighbourhood', 8 = '8-neighbourhood'}

\item{p}{[\code{numerical(1)}]\cr The p value controls the proportion of elements randomly selected to form clusters.}

\item{rescale}{[\code{logical(1)}]\cr If \code{TRUE} (default), the values are rescaled between 0-1.}
}
\value{
Raster with random values ranging from 0-1.
}
\description{
Create a random rectangular cluster neutral landscape model with values ranging 0-1.
}
\examples{
randomClusterNLM(nCol = 10, nRow = 10, neighbourhood = 4, p=0.4)

}
