% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomElement.R
\docType{methods}
\name{randomElementNLM-method}
\alias{randomElementNLM-method}
\alias{randomElementNLM,ANY-method}
\alias{randomElementNLM}
\title{Method randomElementNLM}
\usage{
randomElementNLM(nCol, nRow, n, rescale = TRUE)

\S4method{randomElementNLM}{ANY}(nCol, nRow, n, rescale = TRUE)
}
\arguments{
\item{nCol}{Number of columns for the raster (numerical)}

\item{nRow}{Number of rows for the raster (numerical)}

\item{n}{The number of elements randomly selected to form the basis of nearest-neighbour clusters (numerical)}

\item{rescale}{If \code{TRUE} (Standard), the values are rescaled between 0-1. Otherwise, the distance in raster units is calculated (logical)}
}
\value{
Raster with random values ranging from 0-1.
}
\description{
Method randomElementNLM

Create a random rectangular cluster neutral landscape model with values ranging 0-1.
}
\examples{
\dontrun{
randomElementNLM(nCol = 100, nRow = 100, n = 400)
}

}
