% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random.R
\docType{methods}
\name{randomNLM-method}
\alias{randomNLM-method}
\alias{randomNLM,ANY-method}
\alias{randomNLM}
\title{Method randomNLM}
\usage{
randomNLM(nCol, nRow, rescale = TRUE)

\S4method{randomNLM}{ANY}(nCol, nRow, rescale = TRUE)
}
\arguments{
\item{nCol}{Number of columns for the raster (numerical)}

\item{nRow}{Number of rows for the raster (numerical)}

\item{rescale}{If \code{TRUE} (Standard), the values are rescaled between 0-1. Otherwise, the distance in raster units is calculated (logical)}
}
\value{
Raster with random values ranging from 0-1.
}
\description{
Method randomNLM

Create a spatially random neutral landscape model with values ranging 0-1.
}
\examples{
\dontrun{
randomNLM(nCol = 100, nRow = 100)
}

}
