% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomRectangularCluster.R
\docType{methods}
\name{randomRectangularClusterNLM-method}
\alias{randomRectangularClusterNLM-method}
\alias{randomRectangularClusterNLM,ANY-method}
\alias{randomRectangularClusterNLM}
\title{Method randomRectangularClusterNLM}
\usage{
randomRectangularClusterNLM(nCol, nRow, minL, maxL, rescale = TRUE)

\S4method{randomRectangularClusterNLM}{ANY}(nCol, nRow, minL, maxL,
  rescale = TRUE)
}
\arguments{
\item{nCol}{Number of columns for the raster (numerical)}

\item{nRow}{Number of rows for the raster (numerical)}

\item{minL}{The minimum possible length of width and height for each random rectangular cluster.}

\item{maxL}{The maximum possible length of width and height for each random rectangular cluster.}

\item{rescale}{If \code{TRUE} (Standard), the values are rescaled between 0-1. Otherwise, the distance in raster units is calculated (logical)}
}
\value{
Raster with random values ranging from 0-1.
}
\description{
Method randomRectangularClusterNLM

Create a random rectangular cluster neutral landscape model with values ranging 0-1.
}
\examples{
\dontrun{
randomNLM(nCol = 100, nRow = 100)
}

}
