% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lace.R
\name{reprow}
\alias{reprow}
\title{Repeat rows}
\usage{
reprow(x, n)
}
\arguments{
\item{x}{vector to be repeated}

\item{n}{number of repeats}
}
\description{
This function creates a matrix of a repeated vector where each row is the same.
}
\author{
James R Staley <jrstaley95@gmail.com>
}
