% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescale.R
\docType{methods}
\name{rescaleNLM-method}
\alias{rescaleNLM-method}
\alias{rescaleNLM,RasterLayer-method}
\alias{rescaleNLM}
\title{Method distancegradientNLM}
\usage{
rescaleNLM(r)

\S4method{rescaleNLM}{RasterLayer}(r)
}
\arguments{
\item{r}{RasterLayer containing a neutral landscape}
}
\value{
Rectangular matrix with values ranging from 0-1
}
\description{
Method distancegradientNLM

A rescale in which the values in the raster are linearly rescaled to range between 0 and 1.
}
\examples{
\dontrun{
rescaleNLM(rasterNLM)
}

}
