% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_rescale.R
\name{rescaleNLM}
\alias{rescaleNLM}
\title{rescaleNLM}
\usage{
rescaleNLM(r)
}
\arguments{
\item{r}{[\code{Raster* object}]}
}
\value{
Raster* object with values ranging from 0-1
}
\description{
A rescale in which the values in the raster are linearly rescaled to range between 0 and 1.
}
\examples{

rescaleNLM(nlm_random(10,10))


}
