% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frac_poly_mr.R
\name{summary.frac_poly_mr}
\alias{summary.frac_poly_mr}
\title{Summarizing Fractional Polynomial Fits}
\usage{
\method{summary}{frac_poly_mr}(x, ...)
}
\arguments{
\item{x}{an object of class "frac_poly_mr".}
}
\description{
summary method for class "frac_poly_mr".
}
\author{
James R Staley (js16174@bristol.ac.uk)
}

