% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fracpoly.R
\name{summary.fracpoly_mr}
\alias{summary.fracpoly_mr}
\title{Summary of the fractional polynomial fits}
\usage{
\method{summary}{fracpoly_mr}(x, ...)
}
\arguments{
\item{x}{an object of class "fracpoly_mr".}
}
\description{
summary method for class "fracpoly_mr".
}
\author{
James R Staley <jrstaley95@gmail.com>
}
