% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piecewise.R
\name{summary.piecewise_mr}
\alias{summary.piecewise_mr}
\title{Summary of piecewise_mr}
\usage{
\method{summary}{piecewise_mr}(x, ...)
}
\arguments{
\item{x}{an object of class \code{"piecewise_mr"}}
}
\description{
summary method for class \code{"piecewise_mr"}.
}
\author{
James Staley \href{mailto:jrstaley95@gmail.com}{jrstaley95@gmail.com}
}
