% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_binarize.R
\name{util_binarize}
\alias{util_binarize}
\title{Binarize continuous raster values}
\usage{
util_binarize(nlm_obj, breaks)
}
\arguments{
\item{nlm_obj}{[\code{Raster* object}]}

\item{breaks}{a vector with one or more break percentages}
}
\value{
RasterBrick
}
\description{
Reclassify continuous raster values into binary map cells based upon given
break(s).
}
\details{
Breaks are considerred to be habitat percentages (p). If more than
one percentage is given multiple layers are written in the same brick.
}
\examples{
rndMap <- nlm_random(10, 10)
brks <- c(0.3, 0.5, 0.7)
util_binarize(rndMap, brks)


}
