% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_import_roboto_condensed.R
\name{util_import_roboto_condensed}
\alias{util_import_roboto_condensed}
\title{Import Roboto Condensed font for use in charts}
\usage{
util_import_roboto_condensed()
}
\description{
Roboto Condensed is a trademark of Google.
}
\note{
This will take care of ensuring PDF/PostScript usage. The location of the
font directory is displayed after the base import is complete. It is highly
recommended that you install them on your system the same way you would any
other font you wish to use in other programs.
}
