% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_update_metadata.R
\name{util_update_metadata}
\alias{util_update_metadata}
\title{util_update_metadata}
\usage{
util_update_metadata(x)
}
\arguments{
\item{x}{SpatRaster}
}
\value{
SpatRaster
}
\description{
Internal helper to standardize landscape metadata.
}
\keyword{internal}
