% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_w2cp.R
\name{util_w2cp}
\alias{util_w2cp}
\title{util_w2cp}
\usage{
util_w2cp(weighting)
}
\arguments{
\item{weighting}{A list of numeric values}
}
\value{
Rectangular matrix with values ranging from 0-1
}
\description{
Convert a list of category  weighting  into a 1D array of cumulative proportions.
}
\examples{
util_w2cp(c(0.2, 0.4, 0.6, 0.9))


}
\keyword{internal}
