% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/w2cp.R
\name{w2cp}
\alias{w2cp}
\title{w2cp}
\usage{
w2cp(weighting)
}
\arguments{
\item{weighting}{A list of numeric values}
}
\value{
Rectangular matrix with values ranging from 0-1
}
\description{
Convert a list of category  weighting  into a 1D array of cumulative proportions.
}
\examples{
w2cp(c(0.5, 0.4, 0.6))


}
