% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lace.R
\name{lace}
\alias{lace}
\title{Localised average causal effects}
\usage{
lace(
  y,
  x,
  g,
  covar = NULL,
  q,
  x0q,
  xc_sub = TRUE,
  family = "gaussian",
  xpos = "mean"
)
}
\arguments{
\item{y}{vector of outcome values}

\item{x}{vector of exposure values}

\item{g}{the instrumental variable}

\item{covar}{\code{data.frame} of covariates}

\item{q}{the number of quantiles the exposure distribution is to be split
into within which a causal effect will be fitted, known as localised
average causal effects (LACE) (default: \code{10})}

\item{x0q}{quantiles of x0 (the IV-free exposure)}

\item{xc_sub}{compute the association between the exposure and the insturment
in each quantile of x0}

\item{family}{a description of the error distribution and link function
to be used in the model and is a \code{character} string naming either the
gaussian (i.e. \code{"gaussian"} for continuous data) or binomial
(i.e. \code{"binomial"} for binary data) family function (default: \code{"gaussian"})}

\item{xpos}{the position used to relate \code{x} to the localised average causal
effect, this can either be the mean of the x-values within each quantile
or a percentile (e.g. 0.5 corresponds to the median value)
(default: \code{"mean"})}
}
\value{
\code{lace} returns a \code{data.frame} of the LACE for quantile:

\item{coef}{the LACE in each quantile}

\item{coef_se}{the standard error of the LACE in each quantile}

\item{xmean}{the mean of the exposure in each quantile}

\item{xcoef_sub}{
the association between the exposure and the instrument in each quantile
}

\item{xcoef_sub_se}{
the standard error of the association between the exposure and the
instrument in each quantile
}
}
\description{
\code{lace} computes the localised average causal effect
for each quantile.
}
\author{
James Staley \href{mailto:jrstaley95@gmail.com}{jrstaley95@gmail.com}
}
