% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric_area.R
\name{metric_area}
\alias{metric_area}
\title{metric_area}
\usage{
metric_area(x, poi = NA)
}
\arguments{
\item{x}{[\code{Raster* object}]}

\item{poi}{[\code{numerical}]\cr  Vector of numeric values indicating classes to be reported.}
}
\value{
List of tibbles
}
\description{
Report area for each of the constituent values making up a matrix
}
\details{
Function reports the number of cells and the proportion of total cells made up by each unique class value
}
\examples{
x <- nlm_random(100, 100)
y <- c(0.5, 0.25, 0.25)
z <- util_classify(x, y)

metric_area(z)

}
