% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpd.R
\docType{methods}
\name{mpdNLM-method}
\alias{mpdNLM-method}
\alias{mpdNLM,ANY-method}
\alias{mpdNLM}
\title{Method mpdNLM}
\usage{
mpdNLM(nCol, nRow, h, rescale = TRUE)

\S4method{mpdNLM}{ANY}(nCol, nRow, h, rescale = TRUE)
}
\arguments{
\item{nCol}{Number of columns for the raster (numerical)}

\item{nRow}{Number of rows for the raster (numerical)}

\item{h}{The h value controls the level of spatial autocorrelation in element values.}

\item{rescale}{If \code{TRUE} (Standard), the values are rescaled between 0-1. Otherwise, the distance in raster units is calculated (logical)}
}
\value{
Raster with random values ranging from 0-1.
}
\description{
Method mpdNLM

Create a midpoint displacement neutral landscape model with values ranging 0-1.
}
\examples{
\dontrun{
mpdNLM(nCol = 100, nRow = 100, h = 0.2)
}

}
