% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlm_curds.R
\name{nlm_curds}
\alias{nlm_curds}
\title{nlm_curds}
\usage{
nlm_curds(p, s, resolution = 1)
}
\arguments{
\item{p}{[\code{numerical(x)}]\cr
Vector with percentage(s) to cut of (fill with zeroes)}

\item{s}{[\code{numerical(x)}]\cr
Vector of successive cutting steps (split 1 cell into x cells)}

\item{resolution}{[\code{numerical(1)}]\cr
Resolution of the raster.}
}
\value{
raster
}
\description{
desc.
}
\details{
detail.
}
\examples{
nlm_curds(c(0.25, 0.25, 0.5), c(64, 8, 1))

}
