% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlm_mosaicgibbs.R
\name{nlm_mosaicgibbs}
\alias{nlm_mosaicgibbs}
\title{nlm_mosaicgibbs}
\usage{
nlm_mosaicgibbs(ncol, nrow, resolution = 1, germs, g, R, patch_classes)
}
\arguments{
\item{ncol}{[\code{numerical(1)}]\cr
Number of columns for the raster.}

\item{nrow}{[\code{numerical(1)}]\cr
Number of rows for the raster.}

\item{resolution}{[\code{numerical(1)}]\cr
Resolution of the raster.}

\item{germs}{[\code{numerical(1)}]\cr
Intensity parameter (non-negative integer).}

\item{g}{[\code{numerical(1)}]\cr
Interaction parameter (a value between 0 - hardcore process and 1- poission process).}

\item{R}{[\code{numerical(1)}]\cr
Interaction radius (non-negative integer) for the fitting of the spatial point
pattern process.}

\item{patch_classes}{[\code{numerical(1)}]\cr
Number of classes for germs.}
}
\value{
RasterLayer
}
\description{
Simulate the NLM introduced in Gaucherel (2008).
}
\details{
\code{nlm_mosaicgibbs} offers the second option of simulating a neutral landscape model
described in Gaucherel (2008).
The method works in principal like the tessellation method (\code{nlm_mosaictess}),
but instead of a random point pattern one fits a simulated realization of the Strauss
process. The Strauss process starts with a given number of points and
uses a minimization approach to fit a point pattern with a given interaction
parameter (0 - hardcore process; 1 - poission process) and interaction radius
(distance of points/germs being apart).
}
\examples{
# simulate polygonal landscapes
mosaicgibbs <- nlm_mosaicgibbs(ncol = 40,
                              nrow = 30,
                              germs = 20,
                              g = 0.5,
                              R = 0.02,
                              patch_classes = 12)

\dontrun{
# visualize the NLM
util_plot(mosaicgibbs)
}

}
\references{
Gaucherel, C. (2008) Neutral models for polygonal landscapes with linear
networks. \emph{Ecological Modelling}, 219, 39 - 48.
}
