% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlm_mpd.R
\name{nlm_mpd}
\alias{nlm_mpd}
\title{nlm_mpd}
\usage{
nlm_mpd(nCol, nRow, h, rescale = TRUE)
}
\arguments{
\item{nCol}{[\code{numerical(1)}]\cr Number of columns for the raster.}

\item{nRow}{[\code{numerical(1)}]\cr Number of rows for the raster.}

\item{h}{[\code{numerical(1)}]\cr The h value controls the level of spatial
autocorrelation in element values.}

\item{rescale}{[\code{logical(1)}]\cr If \code{TRUE} (default), the values
are rescaled between 0-1.}
}
\value{
RasterLayer with random values ranging from 0-1.
}
\description{
Create a midpoint displacement neutral landscape model with values ranging 0-1.
}
\examples{
nlm_mpd(nCol = 100, nRow = 100, h = 0.2)


}
