% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlm_percolationmaps.R
\name{nlm_percolation}
\alias{nlm_percolation}
\title{nlm_percolation}
\usage{
nlm_percolation(nCol, nRow, prob = 0.5)
}
\arguments{
\item{nCol}{[\code{numerical(1)}]\cr Number of columns for the raster.}

\item{nRow}{[\code{numerical(1)}]\cr Number of rows for the raster.}

\item{prob}{[\code{numerical(1)}]\cr Probability value for setting a cell either to 0 or 1.}
}
\value{
RasterLayer with random values 0 amd 1.
}
\description{
Create a simple neutral landscape model with either 0 or 1 as values.
}
\examples{
nlm_percolation(nCol = 100, nRow = 100, p=0.5)


}
