% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlm_mosaictess.R
\name{nlm_mosaictess}
\alias{nlm_mosaictess}
\alias{nlm_polylands}
\title{nlm_mosaictess}
\usage{
nlm_mosaictess(ncol, nrow, resolution = 1, germs)
}
\arguments{
\item{ncol}{[\code{numerical(1)}]\cr
Number of columns for the raster.}

\item{nrow}{[\code{numerical(1)}]\cr
Number of rows for the raster.}

\item{resolution}{[\code{numerical(1)}]\cr
Resolution of the raster.}

\item{germs}{[\code{numerical(1)}]\cr
Intensity parameter (non-negative integer).}
}
\value{
RasterLayer
}
\description{
Simulate the NLM introduced in Gaucherel (2008).
}
\details{
\code{nlm_mosaictess} offers the first option of simulating a neutral landscape model
described in Gaucherel (2008). It generates a random point pattern (the germs)
with an independent distribution and uses the voronoi tessellation to simulate
the mosaic landscapes.
}
\examples{
# simulate polygonal landscapes
mosaictess <- nlm_mosaictess(ncol = 30, nrow = 30, germs = 20)

\dontrun{
# visualize the NLM
util_plot(poly_lands)
}

}
\references{
Gaucherel, C. (2008) Neutral models for polygonal landscapes with linear
networks. \emph{Ecological Modelling}, 219, 39 - 48.
}
