% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlm_polylands.R
\name{nlm_polylands}
\alias{nlm_polylands}
\title{nlm_polylands}
\usage{
nlm_polylands(nCol, nRow, resolution = 1, option = 1, germs, g, R,
  patch_classes, rescale = TRUE)
}
\arguments{
\item{nCol}{[\code{numerical(1)}]\cr
Number of columns for the raster.}

\item{nRow}{[\code{numerical(1)}]\cr
Number of rows for the raster.}

\item{resolution}{[\code{numerical(1)}]\cr
Resolution of the raster.}

\item{option}{[\code{numerical(1)}]\cr
If \code{1} (default), the Tessellation method is used to simulate the NLM.
If \code{2}, the Gibbs algorithm method is used to simulate the NLM.}

\item{germs}{[\code{numerical(1)}]\cr
Intensity parameter (a positive number).}

\item{g}{[\code{numerical(1)}]\cr
Interaction parameter (a number between 0 and 1, inclusive - only used when \code{option = 2}).}

\item{R}{[\code{numerical(1)}]\cr
Interaction radius (a non-negative number).}

\item{patch_classes}{[\code{numerical(1)}]\cr
Number of classes for germs.}

\item{rescale}{[\code{logical(1)}]\cr
If \code{TRUE} (default), the values are rescaled between 0-1.
Otherwise, the distance in raster units is calculated.}
}
\value{
RasterLayer
}
\description{
Simulate the NLM introduced in Gaucherel 2008.
}
\details{
The function offers 2 of the 3 NLM described in Gaucheral.
The first one (\code{option = 1}) is a tesselation method.
It generates a random point pattern (the germs) with an indepentend
distribution and uses the voronoi tesselation to simulate the patchy
landscapes.
The second one (\code{option = 2}) is the gibbs alorithm method.
The method works in principal like the tesselation method, but instead of
a random point pattern one fits a simulated realisation of the Strauss
process. The Strauss process starts with a given number of points and
uses a minimization approach to fit a point pattern with a given interaction
parameter (0 - hardcore proces;, 1 - poission process) and interaction radius
(distance of points/germs being apart).
}
\examples{
nlm_polylands(nCol = 50, nRow = 50, germs = 20)

}
\references{
Gaucherel, C. (2008) Neutral models for polygonal landscapes with linear
networks. \emph{Ecological Modelling}, 219, 39 - 48.
}
