% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlm_randomcluster.R
\name{nlm_randomcluster}
\alias{nlm_randomcluster}
\title{nlm_randomcluster}
\usage{
nlm_randomcluster(nCol, nRow, neighbourhood, p, rescale = TRUE)
}
\arguments{
\item{nCol}{[\code{integer(1)}]\cr Number of columns in the raster.}

\item{nRow}{[\code{integer(1)}]\cr Number of rows in the raster.}

\item{neighbourhood}{[\code{numerical(1)}]\cr Clusters are defined using a
set of neighbourhood structures, 4 (Rook's case),
8 (Queen's case).}

\item{p}{[\code{numerical(1)}]\cr The p value controls the proportion of
elements randomly selected to form clusters.}

\item{rescale}{[\code{logical(1)}]\cr If \code{TRUE} (default), the values
are rescaled between 0-1.}
}
\value{
Raster with random values ranging from 0-1.
}
\description{
Create a random cluster nearest-neighbour neutral landscape model with values
ranging 0-1.
}
\examples{
nlm_randomcluster(nCol = 10, nRow = 10, neighbourhood = 8, p = 0.4)

}
