% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NLMR.R
\docType{package}
\name{NLMR-package}
\alias{NLMR}
\alias{NLMR-package}
\title{Simulating Neutral Landscape Models}
\description{
\emph{NLMR} is an R package for simulating neutral landscape models (NLM).
}
\details{
This package contains vignettes that introduce NLM and basic usage
of the \emph{NLMR} package. The vignettes in this package are listed below.

\describe{
\item{\href{https://marcosci.github.io/NLMR/articles/background.html}{NLM
Concepts + Terminology}}{Background information on the concepts and
terminology that underpin the simulation of neutral landscape models.}
\item{\href{https://marcosci.github.io/NLMR/articles/bestiary.html}{NLMR
Bestiary}}{Bestiary of the available neutral landscape models.}
\item{\href{https://marcosci.github.io/NLMR/articles/BasicUsageVignette.html}{
Quickstart Guide}}{Short walk-through of the \emph{NLMR} package and how to
handle the simulations.}
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://marcosci.github.io/NLMR/}
  \item Report bugs at \url{https://github.com/marcosci/NLMR/issues/}
}

}
\author{
\strong{Maintainer}: Marco Sciaini \email{sciaini.marco@gmail.com}

Authors:
\itemize{
  \item Matthias Fritsch \email{matthias.fritsch@forst.uni-goettingen.de}
  \item Craig Simpkins \email{simpkinscraig063@gmail.com}
}

Other contributors:
\itemize{
  \item Cédric Scherer \email{cedricphilippscherer@gmail.com} (Implemented nlm_neigh) [contributor]
}

}
