% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlmr.R
\docType{package}
\name{nlmr-package}
\alias{nlmr}
\alias{nlmr-package}
\title{Simulating Neutral Landscape Models}
\description{
\emph{nlmr} is an R package for simulating neutral landscape models (NLM).
}
\details{
This package contains vignettes that introduce NLM and basic usage
of the \emph{nlmr} package. The vignettes in this package are listed below.

\describe{
\item{\href{https://marcosci.github.io/nlmr/articles/background.html/}{NLM
Concepts + Terminology}}{Background information on the concepts and
terminology that underpin the simulation of neutral landscape models.}
\item{\href{https://marcosci.github.io/nlmr/articles/bestiary.html/}{nlmr
Bestiary}}{Bestiary of the available neutral landscape models.}
\item{\href{https://marcosci.github.io/nlmr/articles/getstarted.html/}{
Quickstart Guide}}{Short walk-through of the \emph{nlmr} package and how to
handle the simulations.}
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://marcosci.github.io/nlmr/}
  \item Report bugs at \url{https://github.com/marcosci/nlmr/issues/}
}

}
\author{
\strong{Maintainer}: Marco Sciaini \email{sciaini.marco@gmail.com}

Authors:
\itemize{
  \item Matthias Fritsch \email{matthias.fritsch@forst.uni-goettingen.de}
}

Other contributors:
\itemize{
  \item Craig Simpkins \email{simpkinscraig063@gmail.com} (Craig reviewed/provided comments to improve the documentation.) [contributor]
  \item Cédric Scherer \email{cedricphilippscherer@gmail.com} (Cédric implemented nlm_neigh) [contributor]
  \item Laura Graham (Laura reviewed the package for rOpenSci, see 
                  https://github.com/ropensci/onboarding/issues/188) [NA]
  \item Jeffrey Hollister (Jeffrey reviewed the package for rOpenSci, see 
                  https://github.com/ropensci/onboarding/issues/188) [NA]
}

}
