% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlmr_plot.R
\name{nlmr_plot}
\alias{nlmr_plot}
\title{nlmr_plot}
\usage{
nlmr_plot(nlm_obj)
}
\arguments{
\item{nlm_obj}{[\code{Raster* object}]}
}
\value{
rasterVis visualization
}
\description{
Plot a Raster* object with the NLMR default theme
}
\examples{
nlm_raster <- randomNLM(10,10)
nlmr_plot(nlm_raster)


}
