% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fracpoly.R
\name{print.fracpoly_mr}
\alias{print.fracpoly_mr}
\title{Print fracpoly_mr}
\usage{
\method{print}{fracpoly_mr}(x, ...)
}
\arguments{
\item{x}{an object of class \code{"fracpoly_mr"}}
}
\description{
print method for class \code{"fracpoly_mr"}.
}
\author{
James Staley \href{mailto:jrstaley95@gmail.com}{jrstaley95@gmail.com}
}
