% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piecewise.R
\name{print.piecewise_mr}
\alias{print.piecewise_mr}
\title{Print piecewise linear fits}
\usage{
\method{print}{piecewise_mr}(x, ...)
}
\arguments{
\item{x}{an object of class "piecewise_mr".}
}
\description{
print method for class "piecewise_mr".
}
\author{
James R Staley <jrstaley95@gmail.com>
}
