% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fracpoly.R
\name{print.summary.fracpoly_mr}
\alias{print.summary.fracpoly_mr}
\title{Print summary of fractional polynomial fits}
\usage{
\method{print}{summary.fracpoly_mr}(x, ...)
}
\arguments{
\item{x}{an object of class "fracpoly_mr".}
}
\description{
print.summary method for class "fracpoly_mr".
}
\author{
James R Staley <js16174@bristol.ac.uk>
}
