% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random.R
\name{randomNLM}
\alias{randomNLM}
\title{randomNML}
\usage{
randomNLM(nCol, nRow, rescale = TRUE)
}
\arguments{
\item{nCol}{Number of columns for the raster (numerical)}

\item{nRow}{Number of rows for the raster (numerical)}

\item{rescale}{If \code{TRUE} (default), the values are rescaled between 0-1.}
}
\value{
Raster with random values ranging from 0-1.
}
\description{
Create a spatially random neutral landscape model with values from a uniform distribution.
}
\details{
The function takes the number of columns and rows as input and creates a RasterLayer with the same extent.
Each raster cell is randomly assigned a value between 0 and 1 drawn from an uniform distribution (\code{runif(1,0,1)}).
If the parameter \code{rescale == TRUE} than the minimum and maximum value are used to rescale the cell values to a range between 0 and 1.
}
\examples{
\dontrun{
randomNLM(nCol = 100, nRow = 100)
}

}
