% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fracpoly.R
\name{summary.fracpoly_mr}
\alias{summary.fracpoly_mr}
\title{Summary of fracpoly_mr}
\usage{
\method{summary}{fracpoly_mr}(x, ...)
}
\arguments{
\item{x}{an object of class \code{"fracpoly_mr"}}
}
\description{
summary method for class \code{"fracpoly_mr"}.
}
\author{
James Staley \href{mailto:jrstaley95@gmail.com}{jrstaley95@gmail.com}
}
