% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piecewise.R
\name{summary.piecewise_mr}
\alias{summary.piecewise_mr}
\title{Summary of piecewise linear fits}
\usage{
\method{summary}{piecewise_mr}(x, ...)
}
\arguments{
\item{x}{an object of class "piecewise_mr".}
}
\description{
summary method for class "piecewise_mr".
}
\author{
James R Staley <js16174@bristol.ac.uk>
}
