% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_calcboundaries.R, R/util_classifymatrix.R
\name{util_calc_boundaries}
\alias{util_calc_boundaries}
\alias{util_classify}
\title{util_calc_boundaries}
\usage{
util_calc_boundaries(x, cumulative_proportions)

util_classify(x, weighting)
}
\arguments{
\item{x}{[\code{matrix(x,y)}]\cr 2D matrix of data values.}

\item{cumulative_proportions}{[\code{numerical}]\cr Vector of class cumulative proportions, as generated by \code{w2cp}.}

\item{weighting}{[\code{numerical}]\cr  Vector of numeric values.}

\item{x}{[\code{matrix(x,y)}]\cr 2D matrix of data values.}
}
\value{
Numerical vector with boundaries for matrix classifcation

Rectangular matrix reclassified values.
}
\description{
Determine upper class boundaries for classification of a matrix with values ranging 0-1 based upon an
vector of cumulative proportions.

Classify a matrix with values ranging 0-1 into proportions based upon a vector of class weightings.
}
\details{
The length of the weighting vector determines the number of classes in the resulting matrix.
}
\examples{
x <- matrix(runif(100,0,1),10,10)
y <- NLMR::w2cp(c(0.5, 0.25, 0.25)) #cumulative proportion
util_calc_boundaries(x,y)

x <- matrix(runif(100, 0, 1), 10, 10)
y <- c(0.5, 0.25, 0.25)
util_classify(x, y)


}
\keyword{internal}
