% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_plot.R
\name{util_plot}
\alias{util_plot}
\title{util_plot}
\usage{
util_plot(nlm_obj, viridis = "A")
}
\arguments{
\item{nlm_obj}{[\code{Raster* object}]}

\item{viridis}{[\code{character}(1)]  Four options are available: "magma" (or "A"), "inferno" (or "B"), "plasma" (or "C"), and "viridis" (or "D", the default option).}
}
\value{
visualization
}
\description{
Plot a Raster* object with the NLMR default theme
}
\examples{
nlm_raster <- nlm_random(10,10)
util_plot(nlm_raster)


}
