% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_plot.R
\name{util_plot}
\alias{util_plot}
\title{util_plot}
\usage{
util_plot(nlm_obj)
}
\arguments{
\item{nlm_obj}{[\code{Raster* object}]}
}
\value{
visualization
}
\description{
Plot a Raster* object with the NLMR default theme
}
\examples{
nlm_raster <- nlm_random(10,10)
util_plot(nlm_raster)


}
