% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_plot.R
\name{util_plot}
\alias{util_plot}
\title{util_plot}
\usage{
util_plot(nlm_obj, scale = "A")
}
\arguments{
\item{nlm_obj}{[\code{Raster* object}]}

\item{scale}{[\code{character}(1)]
Five options are available: "viridis - magma" (= "A"),
                            "viridis - inferno" (= "B"),
                            "viridis - plasma" (= "C"),
                            "viridis - viridis" (= "D", the default option),
                            and "greyscale" (= "E")}
}
\value{
visualization
}
\description{
Plot a Raster* object with the NLMR default theme
}
\examples{
nlm_raster <- nlm_random(10,10)
util_plot(nlm_raster)


}
