% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_rescale.R
\name{util_rescale}
\alias{util_rescale}
\title{util_rescale}
\usage{
util_rescale(r)
}
\arguments{
\item{r}{[\code{Raster* object}]}
}
\value{
Raster* object with values ranging from 0-1
}
\description{
A rescale in which the values in the raster are linearly rescaled to range between 0 and 1.
}
\examples{

# since NLMR uses rescale as the default, one would use it like this
# (or simply set the parameter to \\code{FALSE})
util_rescale(nlm_random(10, 10,  rescale = FALSE))


}
