% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_rescale.R
\name{util_rescale}
\alias{util_rescale}
\title{util_rescale}
\usage{
util_rescale(x)
}
\arguments{
\item{x}{[\code{Raster* object}]}
}
\value{
Raster* object with values ranging from 0-1
}
\description{
Linearly rescale element values in a raster to a range between 0 and 1
}
\details{
Rasters generated by \code{nlm_*} functions are scaled between 0 and 1 as default, this option can be set to \code{FALSE} if needed.
}
\keyword{internal}
