% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/w2cp.R
\docType{methods}
\name{w2cp-method}
\alias{w2cp-method}
\alias{w2cp,ANY-method}
\alias{w2cp}
\title{Method w2cp}
\usage{
w2cp(weighting)

\S4method{w2cp}{ANY}(weighting)
}
\arguments{
\item{weighting}{A list of numeric values}
}
\value{
Rectangular matrix with values ranging from 0-1
}
\description{
Method w2cp

Convert a list of category  weighting  into a 1D array of cumulative proportions.
}
\examples{
\dontrun{
w2cp(rasterNLM)
}

}
